Move the 3 create resources methods into a helper so they can be used to seed the test db
